requirements.txt
# Django Backend Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
transliterate==1.10.2

# Testing Dependencies
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.88.1
factory-boy==3.3.0
coverage==7.3.2

# Development Dependencies
black==23.11.0
flake8==6.1.0